+++ .eslintrc.js
@@ -14,4 +14,7 @@ module.exports = {
       },
     },
   ],
+  parserOptions: {
+    project: './tsconfig.eslint.json',
+  },
 }
