+++ tsconfig.eslint.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["./**/*"],
+  "exclude": ["node_modules"]
+}
